% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_item_order}
\alias{get_item_order}
\title{Internal: extract item order from mirt model or fallback CSV}
\usage{
get_item_order(mod, params_csv_fallback = NULL)
}
\description{
Internal: extract item order from mirt model or fallback CSV
}
\keyword{internal}
